phpPackages: Make phpPackages an attribute of php
This means php packages can easily refer to other php packages by
looking them up in the php attribute. This also means that in
applications where previously both the php package and the
corresponding version of the phpPackages package set had to be
specified, the php package will now suffice, since the phpPackages can
be derived from it.