makeWrapper: Add support for deduplicated vars
This implements the same behaviour as --prefix and --suffix but
removes duplicates of the specified value.

Example usage:

$ makeWrapper $PWD/wrapped $PWD/wrapper --prefix-dedup VAR : abc

$ ./wrapper
Value is abc

$ VAR=aaaa ./wrapper
Value is abc:aaaa

$ VAR=abc ./wrapper
Value is abc

$ VAR=foo:abc ./wrapper
Value is abc:foo

The last example shows how the prefixed value is always put at the
beginning, even when the variable from the outside environment puts it
at the end.

Co-authored-by: Janne Heß <janne@hess.ooo>